finished HelpDialog.tsx
